resumeai_microservice/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
httpx
python-multipart
python-docx
pdfplumber
dotenv
openai
tiktoken
traceback